geoapi/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-gis==1.1
geopandas==1.0.1
idna==3.10
numpy==2.1.1
packaging==24.1
pandas==2.2.3
pyogrio==0.10.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
shapely==2.0.6
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
psycopg2==2.9.9
djangorestframework-simplejwt==5.3.1